fix alert
