+++ cspell-ignore.txt
@@ -8,6 +8,7 @@ Curriculars
 Darkmoon
 Dreamweaver
 maute
+Notif
 Photoshop
 Pooge
 Stroudsburg
