+++ src/components/layouts/MainLayout.tsx
@@ -8,6 +8,8 @@ import {
 } from "@mui/material"
 import { Outlet } from "react-router-dom"
 import { Blob } from "../images/Blob"
+import { Dispatch, useState } from "react"
+import { SuccessNotif } from "../notifs/SuccessNotif"
 
 const MainLayoutRoot = styled("div")(({ theme }) => ({
   backgroundColor: theme.palette.background.paper,
@@ -16,10 +18,20 @@ const MainLayoutRoot = styled("div")(({ theme }) => ({
   position: "fixed",
 }))
 
+export type outletContext = {
+  open: boolean
+  setOpen: Dispatch<React.SetStateAction<boolean>>
+  setMessage: Dispatch<React.SetStateAction<string>>
+}
+
 export const MainLayout = () => {
   const theme = useTheme()
+  const [open, setOpen] = useState(false)
+  const [message, setMessage] = useState("")
+  const alertContext = { open: open, setOpen: setOpen, setMessage: setMessage }
   return (
     <MainLayoutRoot>
+      <SuccessNotif message={message} open={open} setOpen={setOpen} />
       <Link href="/" underline="none">
         <Typography
           variant="h2"
@@ -59,9 +71,10 @@ export const MainLayout = () => {
         <Box
           sx={{
             mask: "linear-gradient(to top, rgba(0,0,0, 1) 0, rgba(0,0,0, 1) 97%, rgba(0,0,0, 0) 100%, rgba(0,0,0, 0) 20%);",
-            overflow: "auto",
+            // overflow: "auto",
             maxHeight: "90vh",
-            zIndex: 2,
+            // zIndex: 2,
+            position: "relative",
             width: "85vw",
             mt: "10vh",
             paddingBottom: 10,
@@ -69,7 +82,7 @@ export const MainLayout = () => {
             scrollbarWidth: "none",
           }}
         >
-          <Outlet />
+          <Outlet context={alertContext} />
         </Box>
       </Container>
     </MainLayoutRoot>
